(Bookmark) Removed several unused methods and variables Corrected all the remaining warnings  * Including deprecated getFontMetrics
Many of these changes made jedit smaller and more ad-hoc.

Notice, this is experimental, so you may wish to revert
some of these changes if a problem arises from it,
although I did test it fairly thoroughly.

This should be bookmarked as a reference for any methods
that were removed that you might need back later.

git-svn-id: https://lateralgm.svn.sourceforge.net/svnroot/lateralgm@257 8f422083-7f27-0410-bc82-93e204be8cd2