Rewrote the implementation of `UpdateSource`. It is simpler and should be a little faster and use less memory.
New abstract class `SetTraverser` is used by `UpdateSource` to store and traverse the set of listeners; one subclass for "hard" listeners and one for "weak" ones. The listeners are stored in a custom singly-linked list, and concurrent modification is not a problem. `add` and `remove` operations use a `HashMap` for fast list node lookup, while traversing uses the linked list directly.

git-svn-id: https://lateralgm.svn.sourceforge.net/svnroot/lateralgm@351 8f422083-7f27-0410-bc82-93e204be8cd2