Reverted BGRA image reading code
the new version in LGM is ~20x slower and I could not determine any difference in function (in particular, the bug that was allegedly fixed by the change can not be reproduced with the older, faster version either, so the real bug must have been fixed elsewhere). Further, prevented a useless raster copy which speeds this up even more.