Improvements
Made reader/writer error messages use JOptionPane and not throw a new
exception, this is what causes the file to tree to fuck up and duplicate
because it never gets to reload performed. Anyway, I changed this
because if you need advanced error messages LGM outputs an error log and
you can also just use Eclipse. Also added shaders and game information
to the GMX writer.