Changed Default Game Icon
The red ball idea is outdated; and I hate the icon I had Josh make with
the tile shits, looks like its done by Visual Studio. Game controller is
the way to go, since its not just an icon for games we need, but
specifically video games, which are primarily played via game pads these
days. Josh will have to vectorize a more unique one, if anybody has any
issues with this I will switch to the ENIGMA logo and that will be the
end of it. Since ENIGMA icon looks great on the window title bar.