Fixed image memory size calculation
The DataBuffer way of how it is suggested via Google reports
astronomical values most likely a result of compression serialization,
and believe me I tested my code thoroughly to make sure. It was also
reporting different sizes based on format which is not what it was
supposed to do because everything becomes BMP when written to ENIGMA
anyway. The new formula for estimating an images memory usage is width *
height * 4 byes per pixel.
Forum post also contains more details.
http://enigma-dev.org/forums/index.php?topic=1708.msg16259#msg16259