Fixed Event Selector
It was missing the <No Key> <Any Key> under keyboard, press, and release
events. Upon a Google search I read several old posts by IsmAvatar about
it being added and found a closed bug report filed on it a while ago,
but could not understand why they were not currently in the event
selector. I have mapped them to a constant Event.EV_NO_KEY and
Event.EV_ANY_KEY as to not just blindly code it with event numbers 0 and
1 since all other keyboard events are mapped to the Java AWT KeyEvent
number.