Fixed GMX Reader/Writer
I was loading the GUID incorrectly, it now reads and writes out
properly.  Also renamed the DPLAY_GUID property to just GAME_GUID which
matches the GMX format, but also because this GUID is not just for
DirectPlay but also for DirectX and a lot of the Windows API, otherwise
you end up with commands being sent to Windows Explorer and stuff, well
according to Microsoft docs of course.