Resource Information Frame
Revamps the resource info frame to use the actual resource and note
require the editing frame to be open, I wrote this rather silly was one
of my first contributions. This change does have the events and moments
realize before the frame opens, objects use saveEvents() which does not
affect the ability to revert changes to the resource after opening the
frame. Also changed it back to use the default font which is monospaced,
and instead derive a smaller version of it.