Fixes behavior of ActionList FRAMES FRAMES was set up to work as a static map of all the open action frames. However, the action list was accessing it in a non-static way by iterating and saving all frames when an instance of an action list is saved. This caused every action list to save every action frame of every other action list, which was redundant.
It is likely more useful to have the map non static to know specifically
the action frames opened only by a single action list. This would allow
you to close the action frames associated with a specific action list, for
example when the object frame closes you can ask the user if they want to
close the open action frames on it.

There is still a reference to all of the open MDI frames so that the MDI
menu in the JMenuBar can do its job.