Transfer handling improvements First, when you use a JVM local DataFlavor, or create one, you need to specify the class. #303 occurred because the JVM was confusing ResNode and EventNode during drag and drop operations as a result of them having the exact same MIME. Later ResourceMenu should be extended to support copying between other ResourceMenus by adding a data flavor for ResourceReference. The Resource in ResNode as well as the Event in EventInstanceNode should both be made serializable so they can be transferred across JVMs.
Second, I have changed the undocked version of the event selector to be a
JDialog. Having it as a JFrame previously, that was always on top, made it
cover dialogs like the new project and unsaved changes dialogs. The
reason I had originally made it a JFrame was so that it would not go
behind the main window when losing focus. I realized the same can be
achieved by setting the parent of the dialog to the main window, and
that it can be made modeless so that it does not block any of the
aforementioned dialogs.