Action List Context Menu Deep Copy (#392)
Make the context menu buttons for the action list do a deep copy instead of a shallow copy of the transferred actions. The `ActionsPaste` method now iterates the action list being transferred and deep copies each of the actions before adding them to the target's model. Also updated the comments of the other deep copy locations for clarity and fixed the indentation of a for loop which was ambiguous. Finally, update the copyright year to extend through 2018 and into 2019.