Safe Varargs Annotation (#397)
Verify that heap pollution is not possible in the vararg methods of the GM stream encoder and decoder. Use the @SafeVarargs annotation to remove the inaccurate compiler warnings. This is safe because the methods assume that elements of the vararg parameters are elements of P, but does not assume the varargs themselves are P[], and thus the elements can be safely used as keys of the correctly typed map.