Non Static Action List Frames (#401)
Change ActionList frames list to non static so each action list remembers only the frames it created. This corrects action lists from saving or closing frames which they did not create. This means that an action list is now required to open an action frame, so orphaned action frames can not be easily created now.