Cleanup GMX Group Writing (#410)
Refactors all of the duplicate group writing methods and extracts a common group writing method where the kind is passed as a parameter. I then refactored all of the recursive methods to use the names of the old group writing methods. I had to tweak each of the recursive methods a bit so they create group/primary elements up front.

Obviously this is quite a few net deletions which should make the GMX writer easier to maintain. The next refactor can actually deduplicate the tree walking altogether and extract each individual resource write method to a helper.