Included File Editing Support (#434)
Completely, brand new, included file editor. The basic decision made here is that we are going with a revamped GameMaker 5 version of data files. This is like GMSv1.4 in that the included file hierarchy is present in the main tree. It differs in that I decided not to open up LGM's main tree to having file extensions in resource names for fear of regression. However, since the new editor, like GM5's, allows you to edit the file name property separately, it should still remain GMSv1.4 compatible.

I had to make a lot of fixes to the file readers and writers. We were ignoring that the data files in GMD are zlib compressed. We were ignoring the store option for GMK 800 and 810 include files. We were throwing out the data files hierarchy in the GMD, which we now reuse for includes. In GMX, we needed an enum mapping for the export action in order to use the property link factory.