Cleanup Enum Codes (#438)
Cleanup code maps using a new helper for generating a code map and a supplemental helper that generates an enum code map. The helpers accept arrays which can contain null keys to allow for gaps in the integer code intervals (e.g, resource kind). I could not use the new helpers for the sprite mask code because some dolt decided to inject EGM polygon mask type in there even though it's not implemented yet and just map it to the same integer as rectangle.