Plugin Preferences (#439)
Completely overhaul the preferences dialog allowing for plugins to have their own preference groups. Each of the existing tabs of preferences have been refactored using the new preference group interface. As a result of the new interface's decoupling of load and save, resetting to the defaults will actually reload the default preferences into the controls now. This change is also backwards compatible with older versions of existing plugins.