Preferences Reload (#440)
This is a minor tweak to the new preferences set up to make it reload each preference group every time the frame is shown. This means if the frame is closed without hitting apply, it will be reloaded before it is next shown. This means the frame no longer has to explicitly load each preference group during its construction.