Generify Undoable Instance/Tile Edit (#468)
Introduce an enum type to the undoable edit class used by tiles and instances. This does not actually utilize the generics, all I did was remove the variable and constructor overload for each undo type and instead replaced them with two generic Object variables. Not only does this make each undo a lot smaller memory-wise it also makes it much easier to expand this to support additional properties in the future by removing ambiguity.