Room Undo Immutable Autoboxing (#490)
Clean up some more of the deprecation warnings by replacing the primitive wrapper constructors with simple autoboxing. Because all primitive wrapper classes are immutable in Java, defensive copying is not necessary. There was no need to use the explicit valueOf methods here because we already have the correct types for autoboxing.